requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
django-redis==5.4.0
redis==5.2.1
celery==5.4.0
drf-spectacular==0.28.0
gunicorn==23.0.0
whitenoise==6.8.2
mypy==1.13.0
django-stubs==5.1.1
djangorestframework-stubs==3.15.1
flake8==7.1.1
black==24.10.0
isort==5.13.2
